+++ components/select/Select.js
@@ -1,14 +1,10 @@
 import React, { PureComponent } from 'react';
 import ReactSelect from 'react-select';
 import PropTypes from 'prop-types';
-import {
-  IconCloseBadgedSmallFilled,
-  IconChevronDownSmallOutline,
-  IconWarningBadgedSmallFilled,
-} from '@teamleader/ui-icons';
+import { IconCloseBadgedSmallFilled, IconChevronDownSmallOutline } from '@teamleader/ui-icons';
 import Box, { omitBoxProps, pickBoxProps } from '../box';
 import Icon from '../icon';
-import { HelpText, TextSmall } from '../../components';
+import { ErrorText, HelpText } from '../../components';
 import { colors } from './constants';
 import theme from './theme.css';
 import cx from 'classnames';
@@ -267,19 +263,6 @@ class Select extends PureComponent {
     );
   };
 
-  renderValidationMessage() {
-    return (
-      <TextSmall className={theme['validation-text']} marginTop={2} display="flex">
-        <Box element="span" className={theme['validation-icon']}>
-          <IconWarningBadgedSmallFilled />
-        </Box>
-        <Box element="span" marginLeft={1}>
-          {this.props.error}
-        </Box>
-      </TextSmall>
-    );
-  }
-
   render() {
     const { components, error, inverse, helpText, size, ...otherProps } = this.props;
 
@@ -304,7 +287,11 @@ class Select extends PureComponent {
           styles={this.getStyles()}
           {...restProps}
         />
-        {error ? this.renderValidationMessage() : helpText && <HelpText inverse={inverse}>{helpText}</HelpText>}
+        {error ? (
+          <ErrorText inverse={inverse}>{error}</ErrorText>
+        ) : (
+          helpText && <HelpText inverse={inverse}>{helpText}</HelpText>
+        )}
       </Box>
     );
   }